fix: main_master_revert (#1)
* Version 444

* Adding Shiboken2 Scripts to Win Build

This fixed the problem with running on windows from a network install. This updated spec worked locally, now testing on github for tomorrow's release.

* Version 445

* Version 446

* Docker: Alpine 3.14, OpenCV 4.5.2, Python 3.9

Suika/opencv-video-minimal has been updated to the latest version of opencv and python. Also updated alipine to 3.14 where more packages are available. Furthermore pyside2 is now the default qt thingy. Basically almsot no crashes should be happening due to strange Qt calls.

* Version 447

* Replacing .md with the cool and improved .yaml

* Incorporated feedback from floo

* Update the bug report issue template

Mostly some grammar and content changes. Removes the preset title, adds instructions to the version field, changes from "older" to "other" in OS options for when new versions come out and we haven't updated the template or people are running betas.

* Fix hydrus client not building for docker

* Version 448

closes #768

* Markup in Fixing_Hydrus_Random_Crashes_Under_Linux.md

* Version 449 (Experimental)

* Update README.md

* Update README.md

* Update README.md

* Version 450

* Version 451

closes #959, closes #955

* Remove unnecessary pypi packages from containers

* Point to the docker package and the "documentation"

* Version 452

* Version 452 hotfix

* Version 453

closes #961, closes #954, closes #904, closes #841

* Version 454

closes #898, closes #876, closes #845

* New Repair Code

This is the first version of significant changes to the database boot repair code and start of module self-repair responsibility. This db should be able to boot with a missing client.caches.db and repair and repopulate almost everything.

* Version 455

closes #975

* Version 456

closes #976, closes #933, closes #922, closes #910

* Version 457

* Version 458

* Version 459

closes #447, closes #982, closes #875, closes #989, closes #986, closes #858, closes #855, closes #807, closes #790

* OpenCV Fix

Let's hope this works for now

* OpenCV Fix Attempt 2

4.5.3.56

* 459 Docker Fix

Another fix, this time for the docker build, fingers crossed.

* Testing new cv builds

Set new cv version in requirements.txt and no longer bundling cv dll in windows (as pyinstaller now does this correct)

* Version 460

closes #996

* Version 461

closes #999, closes #998

* Version 462

closes #1007

* Version 463

* Version 464

closes #1017

* Version 465

closes #761, closes #1023

* Version 466

closes #1027

* Version 467

closes #1028

* Update InnoSetup.iss

* Set Version Info

* Version 468

closes #1036, closes #1039, closes #1033

* Version 469

closes #1042

* Version 470

* Version 470 Hotfix

Fixing a problem with Qt that has type checked signals

* v470 Hotfix 2

Missed it in testing!

* Version 471

* Version 472

closes #770

* Version 472 mpv fix

That's what I get for not doing a test last night. Fingers crossed.

* Version 472 hotfix 2

Ok, undoing mpv for now

* Version 473

Co-authored-by: Hydrus Network Developer <hydrus.admin@gmail.com>
Co-authored-by: suika <2320837+Suika@users.noreply.github.com>
Co-authored-by: Zweibach <shinighost@gmail.com>
Co-authored-by: Paul Friederichsen <floogulinc@gmail.com>
Co-authored-by: bbappserver <bbappserver2@gmail.com>
Co-authored-by: Kaleb Luedtke <trenlymc@gmail.com>
Co-authored-by: Kaleb Luedtke <jluedtk@jci.com>